requirements.txt
streamlit>=1.35,<2
pandas>=2.0,<3
psycopg2-binary>=2.9,<3
